requirements.txt
numpy
pytorch
matplotlib
SALib
scipy
jupyter notebook